Replaced leftover longs

+++ pandaserver/taskbuffer/OraDBProxy.py
@@ -4792,7 +4792,7 @@ class DBProxy:
         timeStart = datetime.datetime.utcnow()
         # check PandaID
         try:
-            long(pandaID)
+            int(pandaID)
         except Exception:
             tmpLog.error("not an integer : %s" % pandaID)
             if getUserInfo:
@@ -6658,7 +6658,7 @@ class DBProxy:
                     if tmpLFN not in retMapLFN[tmpDataset]:
                         retMapLFN[tmpDataset].append(tmpLFN)
                     try:
-                        tmpTotalFileSize += long(tmpFileSize)
+                        tmpTotalFileSize += int(tmpFileSize)
                     except Exception:
                         pass
                 if maxTotalFileSize is None or maxTotalFileSize < tmpTotalFileSize:
@@ -7892,7 +7892,7 @@ class DBProxy:
             self.cur.execute(sql + comment, varMap)
             # get id
             val = self.getvalue_corrector(self.cur.getvalue(varMap[":newID"]))
-            cloudTask.id = long(val)
+            cloudTask.id = int(val)
             # commit
             if not self._commit():
                 raise RuntimeError("Commit error")
@@ -14466,7 +14466,7 @@ class DBProxy:
                 varMap[":current_priority"] = varMap[":priority"]
                 self.cur.execute(sqlT + comment, varMap)
                 val = self.getvalue_corrector(self.cur.getvalue(varMap[":jediTaskID"]))
-                jediTaskID = long(val)
+                jediTaskID = int(val)
                 if properErrorCode:
                     retVal = "succeeded. new jediTaskID={0}".format(jediTaskID)
                 else:
@@ -15312,15 +15312,15 @@ class DBProxy:
             except Exception:
                 nRanges = 8
             try:
-                pandaID = long(pandaID)
+                pandaID = int(pandaID)
             except Exception:
                 pass
             try:
-                jobsetID = long(jobsetID)
+                jobsetID = int(jobsetID)
             except Exception:
                 pass
             try:
-                jediTaskID = long(jediTaskID)
+                jediTaskID = int(jediTaskID)
             except Exception:
                 jediTaskID = None
             iRanges = 0
@@ -15680,11 +15680,11 @@ class DBProxy:
                 try:
                     tmpItems = eventRangeID.split("-")
                     jediTaskID, pandaID, fileID, job_processID, attemptNr = tmpItems
-                    jediTaskID = long(jediTaskID)
-                    pandaID = long(pandaID)
-                    fileID = long(fileID)
-                    job_processID = long(job_processID)
-                    attemptNr = long(attemptNr)
+                    jediTaskID = int(jediTaskID)
+                    pandaID = int(pandaID)
+                    fileID = int(fileID)
+                    job_processID = int(job_processID)
+                    attemptNr = int(attemptNr)
                 except Exception:
                     _logger.error("{0} : wrongly formatted eventRangeID".format(methodName))
                     retList.append(False)
@@ -15784,7 +15784,7 @@ class DBProxy:
                                     zipFileSpec.lfn = eventDict["zipFile"]["lfn"]
                                     zipFileSpec.GUID = str(uuid.uuid4())
                                     if "fsize" in eventDict["zipFile"]:
-                                        zipFileSpec.fsize = long(eventDict["zipFile"]["fsize"])
+                                        zipFileSpec.fsize = int(eventDict["zipFile"]["fsize"])
                                     else:
                                         zipFileSpec.fsize = 0
                                     if "adler32" in eventDict["zipFile"]:
@@ -15804,7 +15804,7 @@ class DBProxy:
                                     varMap[":newRowID"] = self.cur.var(varNUMBER)
                                     self.cur.execute(sqlF + comment, varMap)
                                     val = self.getvalue_corrector(self.cur.getvalue(varMap[":newRowID"]))
-                                    zipRow_ID = long(val)
+                                    zipRow_ID = int(val)
                                     zipRowIdMap[eventDict["zipFile"]["lfn"]] = zipRow_ID
                                     # make an empty file to trigger registration for zip files in Adder
                                     if zipJobSpec.registerEsFiles():
@@ -16578,7 +16578,7 @@ class DBProxy:
                     retI = self.cur.execute(sql1 + comment, varMap)
                     # set PandaID
                     val = self.getvalue_corrector(self.cur.getvalue(varMap[":newPandaID"]))
-                    jobSpec.PandaID = long(val)
+                    jobSpec.PandaID = int(val)
                 else:
                     jobSpec.PandaID = None
                 msgStr = "{0} Generate new PandaID -> {1}#{2} at {3} ".format(
@@ -16630,7 +16630,7 @@ class DBProxy:
                     varMap[":newRowID"] = self.cur.var(varNUMBER)
                     self.cur.execute(sqlFile + comment, varMap)
                     val = self.getvalue_corrector(self.cur.getvalue(varMap[":newRowID"]))
-                    fileSpec.row_ID = long(val)
+                    fileSpec.row_ID = int(val)
                     # change max failure for esmerge
                     if doMerging and fileSpec.type in ["input", "pseudo_input"]:
                         varMap = {}
@@ -20431,7 +20431,7 @@ class DBProxy:
                     tmpLog.debug("skip since HS06sec is None")
                 else:
                     # cap
-                    hs06sec = long(hs06sec)
+                    hs06sec = int(hs06sec)
                     maxHS06sec = 999999999
                     if hs06sec > maxHS06sec:
                         hs06sec = maxHS06sec
@@ -24291,7 +24291,7 @@ class DBProxy:
             retI = self.cur.execute(sql1 + comment, varMap)
             # set PandaID
             val = self.getvalue_corrector(self.cur.getvalue(varMap[":newPandaID"]))
-            jobSpec.PandaID = long(val)
+            jobSpec.PandaID = int(val)
             msgStr = "Generate a fake co-jumbo new PandaID={0} at {1} ".format(jobSpec.PandaID, jobSpec.computingSite)
             tmp_log.debug(msgStr)
             # insert files
@@ -24310,7 +24310,7 @@ class DBProxy:
                 varMap[":newRowID"] = self.cur.var(varNUMBER)
                 self.cur.execute(sqlFile + comment, varMap)
                 val = self.getvalue_corrector(self.cur.getvalue(varMap[":newRowID"]))
-                fileSpec.row_ID = long(val)
+                fileSpec.row_ID = int(val)
             # insert job parameters
             sqlJob = "INSERT INTO ATLAS_PANDA.jobParamsTable (PandaID,jobParameters) VALUES (:PandaID,:param) "
             varMap = {}
